fix last day bug

+++ src/main/java/onefengma/demo/server/model/apibeans/LastRecords.java
@@ -1,10 +1,12 @@
 package onefengma.demo.server.model.apibeans;
 
+import java.math.BigDecimal;
+
 /**
  * Created by chufengma on 16/6/18.
  */
 public class LastRecords {
-    public float count;
-    public float weight;
-    public float sellingMoney;
+    public BigDecimal count;
+    public BigDecimal weight;
+    public BigDecimal sellingMoney;
 }
